move bitset to common crate, move composite file to directory
